+++ src/views/DidEndpointsSign/DidEndpointsSign.module.css
@@ -32,3 +32,9 @@
 .submit {
   composes: buttonPrimary from '../../components/Button/Button.module.css';
 }
+
+.errorTooltip {
+  composes: tooltip from '../../components/ErrorTooltip/ErrorTooltip.module.css';
+  top: 2.5rem;
+  left: 75%;
+}
